Refactor GetPack to new endpoint pattern
